Address review on the shared fixed-range table.
Pad `FixedHashMapCell` so a 12-byte payload becomes 16 bytes and divides the cache line. `computeBucketPrefix` publishes the prefix, so it is no longer const. The bucket-count parameter is `size_t BITS_FOR_BUCKET`.

+++ src/Common/HashTable/HashTable.h
@@ -423,7 +423,7 @@ protected:
     friend class iterator;
     friend class Reader;
 
-    template <typename, typename, typename, typename, typename, typename, Int32, typename>
+    template <typename, typename, typename, typename, typename, typename, size_t, typename>
     friend class TwoLevelHashTable;
 
     template <typename, typename, size_t>
